Separate form submission to its own entrypoint
The form submission will be responded to and then, at the end,
will redirect to the main extension management screen.

This is to make sure that in the final load (the redirect) the
config is ready to be loaded based on the new selection by the
user. Otherwise, if we'd have responded inline in the same manage()
method, the config would've already been loaded, and the user would
have had to refresh to get the enabled/disabled extensions to show
properly.