Move installation to the System model
Call the initial installation (adding the menu item) from the
System model and make sure both ConfigManager and ExtensionManager
call it, so that the menu is available immediately.

Otherwise, it was only called from the controller, which meant that
the user had to visit the management page once before the menu
was available, which was not really very helpful.

Fixes #6